python_server/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
PyJWT==2.9.0
bcrypt==4.2.0
python-multipart==0.0.9
Pillow==11.0.0
regex==2024.9.11
minio==7.2.0
python-dotenv
psycopg2-binary
